rate-service-refactoring: move out utils
